Big rework and tidy-up of ingress
