+++ bpa/docs/storage_subsystem_design.md
@@ -40,7 +40,7 @@ The storage subsystem provides persistent and cached storage for bundles, coordi
 │   BundleStorage     │  │ MetadataStorage │  │   Dispatcher        │
 │   (trait)           │  │ (trait)         │  │                     │
 ├─────────────────────┤  ├─────────────────┤  │  - drop_bundle()    │
-│ - localdisk-storage │  │ - sqlite-storage│  │  - ingest_bundle()  │
+│ - localdisk-storage │  │ - sqlite-storage│  │  - ingress_bundle() │
 │ - bundle_mem        │  │ - metadata_mem  │  │  - poll_waiting()   │
 └─────────────────────┘  └─────────────────┘  └─────────────────────┘
 ```
@@ -253,7 +253,7 @@ Call `bundle_storage_recovery()` to scan all stored bundle data:
 | Condition | Result | Action |
 |-----------|--------|--------|
 | Data + metadata exist | `Valid` | Resume from status checkpoint |
-| Data exists, no metadata | `Orphan` | Insert metadata, re-ingest |
+| Data exists, no metadata | `Orphan` | Full receive pipeline (`process_received_bundle` + `ingress_bundle`) |
 | Duplicate data found | `Duplicate` | Delete spurious copy |
 | Data unparseable | `Junk` | Delete data |
 | Data missing | `Missing` | Skip (race condition) |
@@ -280,7 +280,7 @@ flowchart TD
         ForEach --> Parse["Parse bundle"]
         Parse --> CheckMeta{"Metadata exists?"}
         CheckMeta -->|"Yes"| Confirm["confirm_exists()\nResume from status"]
-        CheckMeta -->|"No"| Insert["Insert metadata\nRe-ingest as orphan"]
+        CheckMeta -->|"No"| Insert["Full receive pipeline\n(orphan)"]
         Confirm --> Mark["Mark metadata as 'confirmed'"]
         Insert --> Mark
     end
