+++ bpa/src/cla/mod.rs
@@ -20,10 +20,6 @@ pub enum Error {
     #[error("The sink is disconnected")]
     Disconnected,
 
-    /// An error occurred while processing a BPv7 bundle.
-    #[error(transparent)]
-    InvalidBundle(#[from] hardy_bpv7::Error),
-
     /// An internal error occurred.
     #[error(transparent)]
     Internal(#[from] Box<dyn core::error::Error + Send + Sync>),
