+++ bpa/src/dispatcher/dispatch.rs
@@ -3,246 +3,6 @@ use futures::{FutureExt, join, select_biased};
 use hardy_bpv7::status_report::ReasonCode;
 
 impl Dispatcher {
-    /// Entry point for bundles received from CLAs.
-    ///
-    /// Parses the CBOR-encoded bundle, validates the format, stores bundle data,
-    /// inserts initial metadata with `New` status, and queues for ingestion.
-    ///
-    /// # Bundle State
-    ///
-    /// - Initial status: `New`
-    /// - Next: `ingest_bundle()` → Ingress filter → `Dispatching`
-    ///
-    /// See [Bundle State Machine Design](../../docs/bundle_state_machine_design.md)
-    /// for the complete state transition diagram.
-    #[cfg_attr(feature = "instrument", instrument(skip_all))]
-    pub async fn receive_bundle(
-        self: &Arc<Self>,
-        mut data: Bytes,
-        ingress_cla: Option<Arc<str>>,
-        ingress_peer_node: Option<hardy_bpv7::eid::NodeId>,
-        ingress_peer_addr: Option<cla::ClaAddress>,
-    ) -> cla::Result<()> {
-        // TODO: Really should not return errors when the bundle content is garbage - it's not the CLAs responsibility to fix it!
-
-        // Count every bundle received from a CLA, before any validation
-        metrics::counter!("bpa.bundle.received").increment(1);
-        metrics::counter!("bpa.bundle.received.bytes").increment(data.len() as u64);
-
-        // Capture received_at as soon as possible
-        let received_at = time::OffsetDateTime::now_utc();
-
-        // Fast pre-check: reject empty, BPv6, and non-CBOR-array data
-        if let Err(e) = crate::cbor::precheck(&data) {
-            metrics::counter!("bpa.bundle.received.dropped").increment(1);
-            return Err(e.into());
-        }
-
-        // Parse the bundle
-        let (bundle, reason, report_unsupported) =
-            match hardy_bpv7::bundle::RewrittenBundle::parse(&data, self.key_provider()) {
-                Err(e) => {
-                    metrics::counter!("bpa.bundle.received.dropped").increment(1);
-                    return Err(e.into());
-                }
-                Ok(hardy_bpv7::bundle::RewrittenBundle::Valid {
-                    bundle,
-                    report_unsupported,
-                }) => (
-                    bundle::Bundle {
-                        metadata: bundle::BundleMetadata {
-                            storage_name: Some(self.store.save_data(data.clone()).await),
-                            read_only: bundle::ReadOnlyMetadata {
-                                received_at,
-                                ingress_peer_node,
-                                ingress_peer_addr,
-                                ingress_cla,
-                                ..Default::default()
-                            },
-                            ..Default::default()
-                        },
-                        bundle,
-                    },
-                    None,
-                    report_unsupported,
-                ),
-                Ok(hardy_bpv7::bundle::RewrittenBundle::Rewritten {
-                    bundle,
-                    new_data,
-                    report_unsupported,
-                    non_canonical: _,
-                }) => {
-                    debug!("Received bundle has been rewritten");
-
-                    data = Bytes::from(new_data);
-                    let storage_name = Some(self.store.save_data(data.clone()).await);
-
-                    (
-                        bundle::Bundle {
-                            metadata: bundle::BundleMetadata {
-                                storage_name,
-                                read_only: bundle::ReadOnlyMetadata {
-                                    received_at,
-                                    ingress_peer_node,
-                                    ingress_peer_addr,
-                                    ingress_cla,
-                                    ..Default::default()
-                                },
-                                ..Default::default()
-                            },
-                            bundle,
-                        },
-                        None,
-                        report_unsupported,
-                    )
-                }
-                Ok(hardy_bpv7::bundle::RewrittenBundle::Invalid {
-                    bundle,
-                    reason,
-                    error,
-                }) => {
-                    debug!("Invalid bundle received: {error}");
-
-                    // Don't bother saving the bundle data, it's garbage
-                    (
-                        bundle::Bundle {
-                            metadata: bundle::BundleMetadata {
-                                read_only: bundle::ReadOnlyMetadata {
-                                    received_at,
-                                    ingress_peer_node,
-                                    ingress_peer_addr,
-                                    ingress_cla,
-                                    ..Default::default()
-                                },
-                                ..Default::default()
-                            },
-                            bundle,
-                        },
-                        Some(reason),
-                        false,
-                    )
-                }
-            };
-
-        if !self.store.insert_metadata(&bundle).await {
-            // Bundle with matching id already exists in the metadata store
-            metrics::counter!("bpa.bundle.received.duplicate").increment(1);
-
-            // TODO: There may be custody transfer signalling that needs to happen here
-
-            // Drop the stored data and do not process further
-            if let Some(storage_name) = &bundle.metadata.storage_name {
-                self.store.delete_data(storage_name).await;
-            }
-            return Ok(());
-        }
-
-        // Report we have received the bundle
-        self.report_bundle_reception(
-            &bundle,
-            if let Some(reason) = &reason {
-                *reason
-            } else if report_unsupported {
-                ReasonCode::BlockUnsupported
-            } else {
-                ReasonCode::NoAdditionalInformation
-            },
-        )
-        .await;
-
-        if reason.is_some() {
-            // Invalid bundle — never entered the pipeline, just clean up
-            self.store.tombstone_metadata(&bundle.bundle.id).await;
-            metrics::counter!("bpa.bundle.received.dropped").increment(1);
-        } else {
-            // Spawn into processing pool for rate limiting
-            self.ingest_bundle(bundle, data).await;
-        }
-        Ok(())
-    }
-
-    /// Spawn bundle ingestion into the processing pool for rate limiting.
-    ///
-    /// This is a rate-limiting wrapper that spawns `ingest_bundle_inner()` into
-    /// the bounded processing pool. The function returns once the task *starts*,
-    /// not when it completes.
-    ///
-    /// # Crash Safety
-    ///
-    /// Because this returns before the Ingress filter completes, bundles remain
-    /// in `New` status until `ingest_bundle_inner()` checkpoints to `Dispatching`.
-    pub(super) async fn ingest_bundle(self: &Arc<Self>, bundle: bundle::Bundle, data: Bytes) {
-        metrics::gauge!("bpa.bundle.status", "state" => crate::otel_metrics::status_label(&bundle.metadata.status)).increment(1.0);
-
-        let dispatcher = self.clone();
-        hardy_async::spawn!(self.processing_pool, "ingest_bundle", async move {
-            dispatcher.ingest_bundle_inner(bundle, data).await
-        })
-        .await;
-    }
-
-    /// Core bundle ingestion logic: validation, Ingress filter, and checkpoint.
-    ///
-    /// # Processing Steps
-    ///
-    /// 1. Validate lifetime (drop if expired)
-    /// 2. Validate hop count (drop if exceeded)
-    /// 3. Execute Ingress filter hook
-    /// 4. Persist any filter mutations (crash-safe ordering)
-    /// 5. **Checkpoint**: Transition status to `Dispatching`
-    /// 6. Call `process_bundle()` for routing decision
-    ///
-    /// # Crash Safety
-    ///
-    /// The checkpoint to `Dispatching` is always persisted after the Ingress
-    /// filter completes. On restart, bundles in `New` status re-run from step 1,
-    /// while bundles in `Dispatching` skip directly to routing.
-    ///
-    /// See [Filter Subsystem Design](../../docs/filter_subsystem_design.md) for
-    /// filter execution details.
-    #[cfg_attr(feature = "instrument", instrument(skip_all,fields(bundle.id = %bundle.bundle.id)))]
-    pub(super) async fn ingest_bundle_inner(&self, mut bundle: bundle::Bundle, mut data: Bytes) {
-        // Ingress filter hook (includes bundle-validity: flags, lifetime, hop-count)
-        (bundle, data) = match self
-            .filter_engine
-            .exec(
-                filter::Hook::Ingress,
-                bundle,
-                data,
-                self.key_provider(),
-                &self.processing_pool,
-            )
-            .await
-        {
-            Ok(filter::ExecResult::Continue(mutation, mut bundle, data)) => {
-                if mutation.data {
-                    if let Some(storage_name) = &bundle.metadata.storage_name {
-                        self.store.replace_data(storage_name, data.clone()).await;
-                    }
-                }
-                // Always checkpoint to Dispatching (crash safety)
-                metrics::gauge!("bpa.bundle.status", "state" => crate::otel_metrics::status_label(&bundle.metadata.status)).decrement(1.0);
-                bundle.metadata.status = bundle::BundleStatus::Dispatching;
-                metrics::gauge!("bpa.bundle.status", "state" => crate::otel_metrics::status_label(&bundle.metadata.status)).increment(1.0);
-                self.store.update_metadata(&bundle).await;
-                (bundle, data)
-            }
-            Ok(filter::ExecResult::Drop(bundle, reason)) => {
-                if let Some(reason) = reason {
-                    return self.drop_bundle(bundle, reason).await;
-                } else {
-                    return self.delete_bundle(bundle).await;
-                }
-            }
-            Err(e) => {
-                error!("Ingress filter execution failed: {e}");
-                return;
-            }
-        };
-
-        self.process_bundle(bundle, data, self.cla_registry()).await;
-    }
-
     /// Queue a bundle for dispatch processing
     pub(super) async fn dispatch_bundle(&self, mut bundle: bundle::Bundle) {
         self.store
@@ -291,7 +51,7 @@ impl Dispatcher {
     ///
     /// See [Routing Design](../../docs/routing_subsystem_design.md) for RIB lookup details.
     #[cfg_attr(feature = "instrument", instrument(skip_all,fields(bundle.id = %bundle.bundle.id)))]
-    async fn process_bundle(
+    pub(super) async fn process_bundle(
         &self,
         mut bundle: bundle::Bundle,
         data: Bytes,
